Refactor a little bit.
