+++ services/indexer/src/events/mod.rs
@@ -1,0 +1,61 @@
+use alloy::primitives::{Address, U256};
+
+pub mod decoders;
+
+#[derive(Debug, Clone)]
+pub struct AccountCreatedEvent {
+    pub leaf_index: U256,
+    pub recovery_address: Address,
+    pub authenticator_addresses: Vec<Address>,
+    pub authenticator_pubkeys: Vec<U256>,
+    pub offchain_signer_commitment: U256,
+}
+
+#[derive(Debug, Clone)]
+pub struct AccountUpdatedEvent {
+    pub leaf_index: U256,
+    pub pubkey_id: u32,
+    pub new_authenticator_pubkey: U256,
+    pub old_authenticator_address: Address,
+    pub new_authenticator_address: Address,
+    pub old_offchain_signer_commitment: U256,
+    pub new_offchain_signer_commitment: U256,
+}
+
+#[derive(Debug, Clone)]
+pub struct AuthenticatorInsertedEvent {
+    pub leaf_index: U256,
+    pub pubkey_id: u32,
+    pub authenticator_address: Address,
+    pub new_authenticator_pubkey: U256,
+    pub old_offchain_signer_commitment: U256,
+    pub new_offchain_signer_commitment: U256,
+}
+
+#[derive(Debug, Clone)]
+pub struct AuthenticatorRemovedEvent {
+    pub leaf_index: U256,
+    pub pubkey_id: u32,
+    pub authenticator_address: Address,
+    pub authenticator_pubkey: U256,
+    pub old_offchain_signer_commitment: U256,
+    pub new_offchain_signer_commitment: U256,
+}
+
+#[derive(Debug, Clone)]
+pub struct AccountRecoveredEvent {
+    pub leaf_index: U256,
+    pub new_authenticator_address: Address,
+    pub new_authenticator_pubkey: U256,
+    pub old_offchain_signer_commitment: U256,
+    pub new_offchain_signer_commitment: U256,
+}
+
+#[derive(Debug, Clone)]
+pub enum RegistryEvent {
+    AccountCreated(AccountCreatedEvent),
+    AccountUpdated(AccountUpdatedEvent),
+    AuthenticatorInserted(AuthenticatorInsertedEvent),
+    AuthenticatorRemoved(AuthenticatorRemovedEvent),
+    AccountRecovered(AccountRecoveredEvent),
+}
